Add authorization

+++ app/services/task/delete_service.py
@@ -1,11 +1,19 @@
+from boto3.dynamodb.conditions import Key
 from botocore.exceptions import ClientError
+from fastapi import HTTPException
 from fastapi.responses import JSONResponse
 
 from app.core.database import dynamodb
 
 
-def delete_service(task_id: str):
+def delete_service(task_id: str, user):
     table = dynamodb.Table("tasks")
+
+    response = table.query(KeyConditionExpression=Key("id").eq(task_id))
+    assigned_user = response["Items"][0].get("assigned_user")
+
+    if assigned_user != user.get("Username"):
+        raise HTTPException(status_code=403, detail="Forbidden")
 
     try:
         response = table.delete_item(
